requirements.txt
apscheduler
requests
beautifulsoup4
discord.py
